Migrate stash_plugin into the pelican CLI format
