+++ cmd/plugin.go
@@ -8,10 +8,11 @@ import (
 	"strings"
 	"time"
 
-	stashcp "github.com/htcondor/osdf-client/v6"
-	"github.com/htcondor/osdf-client/v6/classads"
-	"github.com/htcondor/osdf-client/v6/config"
+	"github.com/pelicanplatform/pelican"
+	"github.com/pelicanplatform/pelican/classads"
+	"github.com/pelicanplatform/pelican/config"
 	log "github.com/sirupsen/logrus"
+	"github.com/spf13/cobra"
 )
 
 var (
@@ -19,6 +20,12 @@ var (
 	commit  = "none"
 	date    = "unknown"
 	builtBy = "unknown"
+
+	// Holds the various plugin commands
+	rootPluginCmd = &cobra.Command{
+		Use:     "plugin",
+		Short:   "Plugin management for HTCSS",
+	}
 )
 
 type Transfer struct {
@@ -26,7 +33,21 @@ type Transfer struct {
 	localFile string
 }
 
-func main() {
+func init() {
+	// Define the file transfer plugin command
+	xferCmd := &cobra.Command{
+		Use:   "transfer",
+		Short: "Run pelican CLI in HTCSS file transfer plugin mode",
+		Args: cobra.ArbitraryArgs,
+		DisableFlagParsing: true, // We have custom flag handling to match HTCSS style.
+		Run: func(_ *cobra.Command, args []string) {stashPluginMain(args)},
+	}
+
+	rootPluginCmd.CompletionOptions.DisableDefaultCmd = true
+	rootPluginCmd.AddCommand(xferCmd)
+}
+
+func stashPluginMain(args []string) {
 	err := config.Init()
 	if err != nil {
 		log.Errorln(err)
@@ -36,76 +57,71 @@ func main() {
 	// Parse command line arguments
 	var upload bool = false
 	// Set the options
-	stashcp.Options.Recursive = false
-	stashcp.Options.ProgressBars = false
-	stashcp.Options.Version = version
-	if err := setLogging(log.PanicLevel); err != nil {
-		log.Panicln("Failed to set log level")
-	}
+	pelican.ObjectClientOptions.Recursive = false
+	pelican.ObjectClientOptions.ProgressBars = false
+	pelican.ObjectClientOptions.Version = version
+	setLogging(log.PanicLevel)
 	methods := []string{"cvmfs", "http"}
 	var infile, outfile, testCachePath string
 	var useOutFile bool = false
 	var getCaches bool = false
 
 	// Pop the executable off the args list
-	_, os.Args = os.Args[0], os.Args[1:]
-	for len(os.Args) > 0 {
+	for len(args) > 0 {
 
-		if os.Args[0] == "-classad" {
+		if args[0] == "-classad" {
 			// Print classad and exit
 			fmt.Println("MultipleFileSupport = true")
 			fmt.Println("PluginVersion = \"" + version + "\"")
 			fmt.Println("PluginType = \"FileTransfer\"")
 			fmt.Println("SupportedMethods = \"stash, osdf\"")
 			os.Exit(0)
-		} else if os.Args[0] == "-version" || os.Args[0] == "-v" {
+		} else if args[0] == "-version" || args[0] == "-v" {
 			fmt.Println("Version:", version)
 			fmt.Println("Build Date:", date)
 			fmt.Println("Build Commit:", commit)
 			fmt.Println("Built By:", builtBy)
 			os.Exit(0)
-		} else if os.Args[0] == "-upload" {
+		} else if args[0] == "-upload" {
 			log.Debugln("Upload detected")
 			upload = true
-		} else if os.Args[0] == "-infile" {
-			infile = os.Args[1]
-			os.Args = os.Args[1:]
+		} else if args[0] == "-infile" {
+			infile = args[1]
+			args = args[1:]
 			log.Debugln("Infile:", infile)
-		} else if os.Args[0] == "-outfile" {
-			outfile = os.Args[1]
-			os.Args = os.Args[1:]
+		} else if args[0] == "-outfile" {
+			outfile = args[1]
+			args = args[1:]
 			useOutFile = true
 			log.Debugln("Outfile:", outfile)
-		} else if os.Args[0] == "-d" {
-			if err := setLogging(log.DebugLevel); err != nil {
-				log.Panicln("Failed to set log level to debug")
-			}
-		} else if os.Args[0] == "-get-caches" {
-			if len(os.Args) < 2 {
+		} else if args[0] == "-d" {
+			setLogging(log.DebugLevel)
+		} else if args[0] == "-get-caches" {
+			if len(args) < 2 {
 				log.Errorln("-get-caches requires an argument")
 				os.Exit(1)
 			}
-			testCachePath = os.Args[1]
-			os.Args = os.Args[1:]
+			testCachePath = args[1]
+			args = args[1:]
 			getCaches = true
-		} else if strings.HasPrefix(os.Args[0], "-") {
-			log.Errorln("Do not understand the option:", os.Args[0])
+		} else if strings.HasPrefix(args[0], "-") {
+			log.Errorln("Do not understand the option:", args[0])
 			os.Exit(1)
 		} else {
 			// Must be the start of a source / destination
 			break
 		}
 		// Pop off the args
-		_, os.Args = os.Args[0], os.Args[1:]
+		args = args[1:]
 	}
 
 	if getCaches {
-		urls, err := stashcp.GetCacheHostnames(testCachePath)
+		urls, err := pelican.GetCacheHostnames(testCachePath)
 		if err != nil {
 			log.Panicln("Failed to get cache URLs:", err)
 		}
 
-		cachesToTry := stashcp.CachesToTry
+		cachesToTry := pelican.CachesToTry
 		if cachesToTry > len(urls) {
 			cachesToTry = len(urls)
 		}
@@ -122,12 +138,12 @@ func main() {
 	//var downloaded int64 = 0
 	var transfers []Transfer
 
-	if len(os.Args) == 0 && (infile == "" || outfile == "") {
+	if len(args) == 0 && (infile == "" || outfile == "") {
 		fmt.Fprint(os.Stderr, "No source or destination specified\n")
 		os.Exit(1)
 	}
 
-	if len(os.Args) == 0 {
+	if len(args) == 0 {
 		// Open the input and output files
 		infileFile, err := os.Open(infile)
 		if err != nil {
@@ -141,8 +157,8 @@ func main() {
 			os.Exit(1)
 		}
 	} else {
-		source = os.Args[:len(os.Args)-1]
-		dest = os.Args[len(os.Args)-1]
+		source = args[:len(args)-1]
+		dest = args[len(args)-1]
 		for _, src := range source {
 			transfers = append(transfers, Transfer{url: src, localFile: dest})
 		}
@@ -156,11 +172,11 @@ func main() {
 		if upload {
 			source = append(source, transfer.localFile)
 			log.Debugln("Uploading:", transfer.localFile, "to", transfer.url)
-			tmpDownloaded, result = stashcp.DoStashCPSingle(transfer.localFile, transfer.url, methods, false)
+			tmpDownloaded, result = pelican.DoStashCPSingle(transfer.localFile, transfer.url, methods, false)
 		} else {
 			source = append(source, transfer.url)
 			log.Debugln("Downloading:", transfer.url, "to", transfer.localFile)
-			tmpDownloaded, result = stashcp.DoStashCPSingle(transfer.url, transfer.localFile, methods, false)
+			tmpDownloaded, result = pelican.DoStashCPSingle(transfer.url, transfer.localFile, methods, false)
 		}
 		startTime := time.Now().Unix()
 		resultAd := classads.NewClassAd()
@@ -182,7 +198,7 @@ func main() {
 			resultAd.Set("TransferTotalBytes", tmpDownloaded)
 		} else {
 			resultAd.Set("TransferSuccess", false)
-			if stashcp.GetErrors() == "" {
+			if pelican.GetErrors() == "" {
 				resultAd.Set("TransferError", result.Error())
 			} else {
 				errMsg := " Failure "
@@ -191,13 +207,13 @@ func main() {
 				} else {
 					errMsg += "downloading "
 				}
-				errMsg += transfer.url + ": " + stashcp.GetErrors()
+				errMsg += transfer.url + ": " + pelican.GetErrors()
 				resultAd.Set("TransferError", errMsg)
-				stashcp.ClearErrors()
+				pelican.ClearErrors()
 			}
 			resultAd.Set("TransferFileBytes", 0)
 			resultAd.Set("TransferTotalBytes", 0)
-			if stashcp.ErrorsRetryable() {
+			if pelican.ErrorsRetryable() {
 				resultAd.Set("TransferRetryable", true)
 				retryable = true
 			} else {
@@ -241,15 +257,6 @@ func main() {
 	} else {
 		os.Exit(1)
 	}
-}
-
-func setLogging(logLevel log.Level) error {
-	textFormatter := log.TextFormatter{}
-	textFormatter.DisableLevelTruncation = true
-	textFormatter.FullTimestamp = true
-	log.SetFormatter(&textFormatter)
-	log.SetLevel(logLevel)
-	return nil
 }
 
 // readMultiTransfers reads the transfers from a Reader, such as stdin
