+++ cmd/plugin_test.go
