+++ errorAccum_test.go
@@ -1,4 +1,4 @@
-package stashcp
+package pelican
 
 import (
 	"errors"
