+++ resources/10-stash-plugin.conf
