+++ resources/update-namespaces.sh
